src/requirements.txt
Flask
Flask-Cors
GitPython
PyYAML
streamlit